refactor(api): break combination and template types into separate files
